rename motion profile classes and move into lib package
(cherry picked from commit 65d65d4)
