+++ src/org/team1540/kingbass/subsystems/DriveTrain.java
@@ -18,7 +18,7 @@ import static org.team1540.kingbass.Tuning.rReverseSensor;
 import com.ctre.CANTalon;
 import edu.wpi.first.wpilibj.command.Subsystem;
 import org.team1540.kingbass.commands.drivetrain.JoystickDrive;
-import org.team1540.kingbass.motion.MotionProfile;
+import org.team1540.lib.motionprofile.ProfileExecuter;
 
 /**
  * 6-cim drive train.
@@ -40,8 +40,8 @@ public class DriveTrain extends Subsystem {
   private CANTalon[] lSlaves = {lSlaveA, lSlaveB};
   private CANTalon[] rSlaves = {rSlaveA, rSlaveB};
 
-  private MotionProfile leftProfile;
-  private MotionProfile rightProfile;
+  private ProfileExecuter leftProfile;
+  private ProfileExecuter rightProfile;
 
   private int driveDirection = 1;
 
@@ -119,8 +119,8 @@ public class DriveTrain extends Subsystem {
   }
 
   public void setMp(double[][] left, double[][] right) {
-    leftProfile = new MotionProfile(lMain, left, left.length);
-    rightProfile = new MotionProfile(rMain, right, right.length);
+    leftProfile = new ProfileExecuter(lMain, left, left.length);
+    rightProfile = new ProfileExecuter(rMain, right, right.length);
   }
 
   public void setPID(double p, double i, double d, double f) {
